backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
pillow>=12.1.0
rembg[cpu]>=2.0.75
onnxruntime>=1.16.0
pydantic==2.5.0